test(OP#9528): Refactor tests to own module
